langgraph_app/requirements.txt
# === Core Framework ===
fastapi==0.120.4
uvicorn[standard]==0.37.0
pydantic==2.12.3
asyncpg

# === LangGraph & LangChain ===
langgraph==1.0.2
langchain==1.0.3
langchain-openai==1.0.1

# === Model APIs ===
openai==2.7.0
anthropic>=0.8.0
google-generativeai>=0.5.0

# === Database & Persistence ===
sqlalchemy>=2.0.0
psycopg2-binary==2.9.9

# === Background Tasks / Queues ===
celery>=5.3.0
redis
python-frontmatter

# === Networking & HTTP ===
requests>=2.31.0
httpx==0.28.0
beautifulsoup4>=4.12.0

# === Publishing Integrations (optional) ===
linkedin-api==2.0.1
tweepy>=4.14.0

# === Content Processing ===
Pillow>=10.0.0
markdown>=3.4.0
html2text==2020.1.16
python-magic==0.4.27

# === Config & Serialization ===
python-dotenv==0.21.1
pyyaml==6.0.2
marshmallow>=3.20.0
jsonschema>=4.0.0
slowapi>=0.1.9
prometheus-client>=0.19.0

# === Quality & NLP Utilities ===
textstat>=0.7.3
pyspellchecker>=0.7.3
language-tool-python>=2.7.0

# === Validation & Multipart ===
python-multipart>=0.0.6

# === Utilities ===
python-dateutil>=2.8.0
pytz>=2023.3
click>=8.0.0
rich>=13.0.0
structlog>=23.0.0

# === Security ===
cryptography>=41.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# === Optional Email / File Exports ===
sendgrid>=6.10.0
openpyxl>=3.1.0
python-docx==0.8.11
reportlab>=4.0.0
